3 ms (78.42%) 56.6 MB (48.94%) 11-17-2025 0643-maximum-average-subarray-i Easy .java

+++ LeetCode/Easy/0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,12 +1,12 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
         // Initialize sum to store the sum of the first 'k' elements
-        double sum = 0;
+        long sum = 0;
         
         for(int i=0; i<k; i++) sum += nums[i];
 
         // Initialize max with the sum of the first window        
-        double max = sum;
+        long max = sum;
         
         // Slide the window from index k to the end of the array
         for(int i=k; i<nums.length; i++){
